refactor test plugins to cover the main scenarios
