+++ test-plugins/PluginAll/FalcoSecurity.Plugin.Sdk.Test.PluginAllCaps.csproj
@@ -13,7 +13,7 @@
 		<PackageReference Include="DNNE" Version="1.0.32" />
 	</ItemGroup>
 	<ItemGroup>
-		<ProjectReference Include="..\FalcoSecurity.Plugin.Sdk.Generators\FalcoSecurity.Plugin.Sdk.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
-		<ProjectReference Include="..\FalcoSecurity.Plugin.Sdk\FalcoSecurity.Plugin.Sdk.csproj" />
+		<ProjectReference Include="..\..\FalcoSecurity.Plugin.Sdk\FalcoSecurity.Plugin.Sdk.csproj" />
+		<ProjectReference Include="..\..\FalcoSecurity.Plugin.Sdk.Generators\FalcoSecurity.Plugin.Sdk.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
 	</ItemGroup>
 </Project>
