+++ test-plugins/PluginEventSourceOnly/FalcoSecurity.Plugin.Sdk.Test.PluginEventSourceOnly.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<TargetFramework>net6.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+		<EnableDynamicLoading>true</EnableDynamicLoading>
+		<DnneAddGeneratedBinaryToProject>true</DnneAddGeneratedBinaryToProject>
+		<DnneNativeBinaryName>plugin_native</DnneNativeBinaryName>
+		<PublishSingleFile>true</PublishSingleFile>
+	</PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="DNNE" Version="1.0.32" />
+	</ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\..\FalcoSecurity.Plugin.Sdk\FalcoSecurity.Plugin.Sdk.csproj" />
+		<ProjectReference Include="..\..\FalcoSecurity.Plugin.Sdk.Generators\FalcoSecurity.Plugin.Sdk.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
+	</ItemGroup>
+</Project>
