+++ test-plugins/PluginEventSourceOnly/Plugin.cs
@@ -5,24 +5,17 @@ namespace FalcoSecurity.Plugin.Sdk.Test
 {
     [FalcoPlugin(
         Id = 999,
-        Name = "test_plugin",
-        Description = "test_description_string",
-        Contacts = "me",
+        Name = "test.event_source_only",
+        Description = "test description!",
+        Contacts = "<test test@test.com>",
         RequiredApiVersion = "2.0.0",
         Version = "1.2.3")]
-    public class TestPlugin : PluginBase, IEventSource, IFieldExtractor		
+    public class TestPlugin : PluginBase, IEventSource		
     {
         public string EventSourceName => "test_eventsource";
 
         public IEnumerable<OpenParam> OpenParameters
             => Enumerable.Empty<OpenParam>();
-
-        public IEnumerable<ExtractionField> Fields
-            => Enumerable.Empty<ExtractionField>();
-
-
-        public IEnumerable<string> EventSourcesToExtract
-            => Enumerable.Empty<string>();
 
         public IEventSourceInstance Open(IEnumerable<OpenParam>? openParams)
         {
@@ -32,11 +25,6 @@ namespace FalcoSecurity.Plugin.Sdk.Test
         public void Close(IEventSourceInstance instance)
         {
             instance.Dispose();
-        }
-
-        public void Extract(IExtractionRequest extraction, IEventReader evt)
-        {
-            throw new NotImplementedException();
         }
     }
 }
